fix: fixing bug in modeldowloader that cannot pull model
